Various fixes

+++ app/src/main/java/spit/matrix2017/Activities/EventDetails.java
@@ -22,6 +22,7 @@ import android.support.design.widget.CollapsingToolbarLayout;
 import android.support.design.widget.FloatingActionButton;
 import android.support.v4.app.ActivityCompat;
 import android.support.v4.content.ContextCompat;
+import android.support.v4.view.animation.FastOutLinearInInterpolator;
 import android.support.v4.view.animation.LinearOutSlowInInterpolator;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.app.AppCompatActivity;
@@ -35,6 +36,8 @@ import android.view.Gravity;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
+import android.view.animation.LinearInterpolator;
+import android.view.animation.PathInterpolator;
 import android.widget.ImageButton;
 import android.widget.ImageView;
 import android.widget.TextView;
@@ -95,8 +98,8 @@ public class EventDetails
         setPrizes(getIntent().getStringExtra("prizes"));
         setContacts(getIntent().getStringExtra("contact1name"), getIntent().getStringExtra("contact1no"), getIntent().getStringExtra("contact2name"), getIntent().getStringExtra("contact2no"));
 
-        isFavouriteEvent = getIntent().getLongExtra("favorite", 0) == 1;
-        isReminderSet = getIntent().getLongExtra("reminder", 0) == 1;
+        isFavouriteEvent = getIntent().getIntExtra("favorite", 0) == 1;
+        isReminderSet = getIntent().getIntExtra("reminder", 0) == 1;
 
         mainImageView = (ImageView) findViewById(R.id.main_imageView);
         assert mainImageView != null;
